feat(port): AnglesiteBridgeCore split — webview-agnostic message schema
Phase 2 of the cross-platform port (#567), step 1 of 3 per the design doc's
phasing ("AnglesiteBridgeCore split → PodmanSiteRuntime → Adwaita shell").

The `anglesite` script-message schema, decoding, and routing logic — previously
living inside `AnglesiteScriptHandler` (a `WKScriptMessageHandler`, so the whole
file needed WebKit) — moves into a new portable `AnglesiteBridgeCore` target with
no WebKit import. Each platform's webview adapter (`AnglesiteBridge`/WKWebView
today; WebKitGTK/WebView2 for the Linux/Windows shells later) becomes a thin
wrapper: unwrap the native message body, call the shared dispatcher, apply the
platform-specific reply mechanism.

## Changes

- **New `AnglesiteBridgeCore` target**: `AnglesiteMessageDispatcher` (the
  `dispatch(body:via:onVisibleElements:onCanvasSelection:onComputedStyles:)`
  static method + `DispatchResult` enum + handler typealiases + the
  `scriptMessageNamespace = "anglesite"` constant every platform shares) and
  `AnglesiteOverlayBundle` (portable overlay.js lookup — locates + reads the
  compiled edit-overlay source from a `Bundle`; each platform adapter wraps the
  resulting `String` in its own native script-injection type).
- **`AnglesiteScriptHandler`** (stays in `AnglesiteBridge`, WKWebView-specific):
  keeps its exact public API (constructor, `dispatch`, the deprecated `handle`
  shim, `VisibleElementsHandler`/`CanvasSelectionHandler`/`ComputedStylesHandler`
  typealiases) — all now forwarding to `AnglesiteMessageDispatcher` — so the
  three existing app-target call sites (`PreviewView`, `ComponentEditorView`)
  need zero changes. Its own remaining job is exactly what `WKScriptMessage`
  requires: unwrap `message.body`/`.webView`, evaluate the reply script.
- **`WebViewBridge`**: `scriptMessageNamespace` and `makeOverlayUserScript(in:)`
  now source from the portable core instead of duplicating the logic.
- **Tests**: `AnglesiteScriptHandlerTests.swift`'s dispatch-logic cases move to
  `Tests/AnglesiteBridgeCoreTests/AnglesiteMessageDispatcherTests.swift`
  (renamed to test `AnglesiteMessageDispatcher` directly) — they now run on
  every platform, per the design doc's "existing AnglesiteBridgeTests move to
  the core target and run on all platforms." The two `handle()` deprecated-shim
  tests stay in `AnglesiteBridgeTests`, since that shim only exists on the
  WKWebView-specific class.
- **Package.swift**: new `AnglesiteBridgeCore` target + library product;
  `AnglesiteBridge` depends on it; `AnglesiteBridgeCoreTests` test target added;
  both new targets join the off-Darwin `portableTargets` set.

## Verification

On the Linux dev box: clean `swift build` + `swift test` — 35 tests, 5 suites,
0 failures (26 pre-existing + 9 new `AnglesiteMessageDispatcherTests`).

Could not verify the Darwin-only `AnglesiteBridge`/`AnglesiteApp` targets
compile — this sandbox has no Xcode/WebKit. Reviewed `AnglesiteScriptHandler.swift`/
`WebViewBridge.swift` by hand for API-compatibility with their three known
call sites (`PreviewView.swift`, `ComponentEditorView.swift`) and confirmed no
other file references the removed inline `DispatchResult`/`dispatch`
implementations. CI's macOS legs (`build-test`, Xcode project sync) will be the
first real compile of this on Darwin — flagging that explicitly for review
rather than claiming full verification.
