+++ Sources/AnglesiteBridge/AnglesiteScriptHandler.swift
@@ -1,36 +1,30 @@
 import Foundation
 import WebKit
 import AnglesiteCore
+import AnglesiteBridgeCore
 
-/// `WKScriptMessageHandler` for the `anglesite` namespace. Four message types ride this bridge:
-///
-/// 1. `anglesite:apply-edit` (an `EditMessage`) — routed through the injected `EditRouter`,
-///    reply delivered back to the WKWebView via `window.anglesite?._handleReply?.(<reply>)`.
-/// 2. `anglesite:visible-elements` (a `VisibleElementReport`, #145/B.1) — dispatched to the
-///    optional `onVisibleElements` callback. No reply; the JS side doesn't await one.
-/// 3. `anglesite:canvas-selection` (a `CanvasSelectionMessage`) — dispatched to the optional
-///    `onCanvasSelection` callback. Posted by the component-harness canvas overlay
-///    (`JS/edit-overlay/src/component-canvas.ts`) on click. No reply.
-/// 4. `anglesite:computed-styles` (a `ComputedStylesReport`) — dispatched to the optional
-///    `onComputedStyles` callback, posted alongside a canvas selection. No reply.
-///
-/// The interesting logic lives in `dispatch(body:via:onVisibleElements:onCanvasSelection:onComputedStyles:)`
-/// — it's unit-tested independent of `WKScriptMessage`, which has no public initializer and is
-/// awkward to fake.
+/// `WKScriptMessageHandler` adapter for the `anglesite` namespace — the WKWebView-specific thin
+/// layer over ``AnglesiteMessageDispatcher`` (cross-platform port design §6 "AnglesiteBridgeCore
+/// split"). All the message schema, decoding, and routing logic lives in the portable core; this
+/// class's own job is exactly two things `WKScriptMessage` requires: unwrap `message.body`/
+/// `.webView`, and evaluate the reply script back into the page.
 ///
 /// **API change vs prior versions:** the primary entry point is now
-/// `dispatch(body:via:onVisibleElements:onCanvasSelection:onComputedStyles:)`. All current
-/// callers are internal to this repo (the `WKScriptMessageHandler` impl below, the unit tests,
-/// and `PreviewView`'s production init) and use `dispatch` directly. The old `handle(body:via:)`
+/// `dispatch(body:via:onVisibleElements:onCanvasSelection:onComputedStyles:)` (forwarding to
+/// ``AnglesiteMessageDispatcher/dispatch(body:via:onVisibleElements:onCanvasSelection:onComputedStyles:)``).
+/// All current callers are internal to this repo (the `WKScriptMessageHandler` impl below, the
+/// unit tests — now in `AnglesiteBridgeCoreTests`, testing `AnglesiteMessageDispatcher` directly
+/// — and `PreviewView`'s production init) and use `dispatch` directly. The old `handle(body:via:)`
 /// signature is kept as a `@available(*, deprecated)` shim below — apply-edit only, matching its
 /// prior behavior — so a downstream framework consumer hitting this gets a fix-it instead of a
 /// cryptic missing-member error. The migration is mechanical: rename `handle` → `dispatch`, pass
 /// `nil` for the optional handlers to preserve the apply-edit-only behavior, and match on the
 /// richer `DispatchResult` enum instead of `Result<EditReply, EditMessage.DecodeError>`.
 public final class AnglesiteScriptHandler: NSObject, WKScriptMessageHandler {
-    public typealias VisibleElementsHandler = @Sendable ([VisibleElement]) async -> Void
-    public typealias CanvasSelectionHandler = @Sendable (CanvasSelectionMessage) async -> Void
-    public typealias ComputedStylesHandler = @Sendable (ComputedStylesReport) async -> Void
+    public typealias VisibleElementsHandler = AnglesiteMessageDispatcher.VisibleElementsHandler
+    public typealias CanvasSelectionHandler = AnglesiteMessageDispatcher.CanvasSelectionHandler
+    public typealias ComputedStylesHandler = AnglesiteMessageDispatcher.ComputedStylesHandler
+    public typealias DispatchResult = AnglesiteMessageDispatcher.DispatchResult
 
     private let router: EditRouter
     private let onVisibleElements: VisibleElementsHandler?
@@ -53,41 +47,9 @@ public final class AnglesiteScriptHandler: NSObject, WKScriptMessageHandler {
         super.init()
     }
 
-    /// Outcome of dispatching one incoming message body. The script handler's
-    /// `userContentController` reads this to decide whether to emit a reply or log a rejection.
-    public enum DispatchResult: Sendable {
-        /// `anglesite:apply-edit` succeeded; emit the reply back to the WKWebView.
-        case editReply(EditReply)
-        /// `anglesite:visible-elements` was forwarded to the optional handler.
-        case visibleElementsHandled
-        /// `anglesite:visible-elements` arrived but no `onVisibleElements` handler is installed.
-        /// Useful for tests; in production the wiring is checked at handler-init time.
-        case visibleElementsDropped
-        /// `anglesite:canvas-selection` was forwarded to the optional handler.
-        case canvasSelectionHandled
-        /// `anglesite:canvas-selection` arrived but no `onCanvasSelection` handler is installed.
-        case canvasSelectionDropped
-        /// `anglesite:computed-styles` was forwarded to the optional handler.
-        case computedStylesHandled
-        /// `anglesite:computed-styles` arrived but no `onComputedStyles` handler is installed.
-        case computedStylesDropped
-        /// Body was undecodable. Log and move on.
-        case rejected(RejectionReason)
-
-        public enum RejectionReason: Sendable, Equatable {
-            case notAnObject
-            case missingType
-            case wrongType
-            case unknownType(String)
-            case editDecode(EditMessage.DecodeError)
-            case visibleElementsDecode(VisibleElementReport.DecodeError)
-            case canvasSelectionDecode(ComponentCanvasDecodeError)
-            case computedStylesDecode(ComponentCanvasDecodeError)
-        }
-    }
-
-    /// Peek at the `type` field, dispatch to the matching decoder, and route. Pure — no I/O
-    /// beyond the router call and the visible-elements handler call.
+    /// Forwards to ``AnglesiteMessageDispatcher/dispatch(body:via:onVisibleElements:onCanvasSelection:onComputedStyles:)``
+    /// — kept here so existing call sites (this class's own `userContentController`, and any
+    /// code written against the pre-split API) don't need to change.
     public static func dispatch(
         body: Any,
         via router: EditRouter,
@@ -95,53 +57,13 @@ public final class AnglesiteScriptHandler: NSObject, WKScriptMessageHandler {
         onCanvasSelection: CanvasSelectionHandler? = nil,
         onComputedStyles: ComputedStylesHandler? = nil
     ) async -> DispatchResult {
-        guard let dict = body as? [String: Any] else { return .rejected(.notAnObject) }
-        guard let rawType = dict["type"] else { return .rejected(.missingType) }
-        guard let typeStr = rawType as? String else { return .rejected(.wrongType) }
-
-        switch typeStr {
-        case EditMessage.MessageType.applyEdit.rawValue:
-            switch EditMessage.decode(from: body) {
-            case .success(let message):
-                let reply = await router.apply(message)
-                return .editReply(reply)
-            case .failure(let error):
-                return .rejected(.editDecode(error))
-            }
-
-        case VisibleElementReport.messageType:
-            switch VisibleElementReport.decode(from: body) {
-            case .success(let report):
-                guard let handler = onVisibleElements else { return .visibleElementsDropped }
-                await handler(report.elements)
-                return .visibleElementsHandled
-            case .failure(let error):
-                return .rejected(.visibleElementsDecode(error))
-            }
-
-        case CanvasSelectionMessage.messageType:
-            switch CanvasSelectionMessage.decode(from: body) {
-            case .success(let message):
-                guard let handler = onCanvasSelection else { return .canvasSelectionDropped }
-                await handler(message)
-                return .canvasSelectionHandled
-            case .failure(let error):
-                return .rejected(.canvasSelectionDecode(error))
-            }
-
-        case ComputedStylesReport.messageType:
-            switch ComputedStylesReport.decode(from: body) {
-            case .success(let report):
-                guard let handler = onComputedStyles else { return .computedStylesDropped }
-                await handler(report)
-                return .computedStylesHandled
-            case .failure(let error):
-                return .rejected(.computedStylesDecode(error))
-            }
-
-        default:
-            return .rejected(.unknownType(typeStr))
-        }
+        await AnglesiteMessageDispatcher.dispatch(
+            body: body,
+            via: router,
+            onVisibleElements: onVisibleElements,
+            onCanvasSelection: onCanvasSelection,
+            onComputedStyles: onComputedStyles
+        )
     }
 
     /// Deprecated forwarder for the old `handle(body:via:)` signature so out-of-tree adopters
